freebsd-update(8): Lock working directory while in use
Running multiple instances of freebsd-update at the same time on the
same working directory often results in a failed update or worse.  This is
especially problematic for users who use freebsd-update in cron to check
for updates. If freebsd-update is run automatically while another instance
is already in progress, the resulting failure can be difficult to diagnose.

This update prevents the collision by locking the working directory and
failing to start an update if the working directory is already locked.

PR: 266270